bug fix - compilation error for  64-bit systems where size_t was implicitely converted to uint32

+++ NihavGameEngine/source/Library.Desktop/Hashmap.inl
@@ -21,7 +21,7 @@ namespace Library
 
 	template<typename TKey, typename TData, typename HashFunctor>
 	Hashmap<TKey, TData, HashFunctor>::Hashmap(std::initializer_list<typename Hashmap::PairType> initList) :
-		buckets(initList.size() , true), mSize(0), mLowestUsedBucketIndex(initList.size() - 1), mHighestUsedBucketIndex(0)
+		buckets(static_cast<std::uint32_t>(initList.size()) , true), mSize(0), mLowestUsedBucketIndex(static_cast<std::uint32_t>(initList.size()) - 1), mHighestUsedBucketIndex(0)
 	{
 		for (auto& pair : initList)
 		{
